// types.d.ts
export type MediaType = "movie" | "tv";
// ...
export interface MediaListQuery {
  type:
    | "movies_popular"
    | "movies_now-playing"
    | "movies_top-rated"
    | "movies_upcoming"
    | "tv-shows_popular"
    | "tv-shows_on-the-air"
    | "tv-shows_top-rated"
    | "tv-shows_airing-today";

  page: number;
}